Ensure dtype consistency in meter.update() calls

+++ dllm/core/trainers/bd3lm.py
@@ -206,7 +206,7 @@ class BD3LMTrainer(MDLMTrainer):
             self.meter.update(
                 split="train" if model.training else "eval",
                 value=torch.zeros_like(maskable_mask, dtype=logits.dtype),
-                weight=maskable_mask.detach(),
+                weight=maskable_mask.to(dtype=logits.dtype).detach(),
             )
             return (zero, outputs) if return_outputs else zero
 
@@ -232,7 +232,7 @@ class BD3LMTrainer(MDLMTrainer):
         self.meter.update(
             split="train" if model.training else "eval",
             value=token_nll.detach(),
-            weight=maskable_mask.detach(),
+            weight=maskable_mask.to(dtype=logits.dtype).detach(),
         )
 
         # === 7. Normalize loss ===
